remove the old graph interpreter cleanup the new graph_interpreter and patronus lowering
