+++ protocols/src/ir/mod.rs
@@ -1,6 +1,5 @@
 pub mod edge_contract;
 pub mod graph_interpreter;
-pub mod new_graph_interpreter;
 pub mod graphviz;
 pub mod lowering;
 mod proto_graph;
